requirements.txt
django
tensorflow
pillow